API changes and move example to test folder
Having a global components table allowed most of the entity’s functions
to be defined outside of the entity table. This allowed each entity to
only need a reference to the functions instead of each entity needing
its own functions. However, by each entity having its own table of
components, the ID no longer needs to be accessible outside of the
entity. Removing the `getID` function means the only accessor is
`getComponent`. The API can then be simplified to a single
`Entity:get(component)` function. To check if an entity has a
component, a new `Entity:has(component)` has been added.

The `entities` table is changed to store the reference to the entity as
the index, meaning that traversing a table returned by
`ecs.getEntities(component)` with a for loop no longer needs both the
key and value variables.